+++ main/automap.c
@@ -1308,7 +1308,7 @@ void DrawAllEdges ()
 //finds edge, filling in edge_ptr. if found old edge, returns index, else return -1
 static int AutomapFindEdge (int v0,int v1,Edge_info **edge_ptr)
 {
-	long vv, evv;
+	int vv, evv;
 	int hash,oldhash;
 	int ret, ev0, ev1;
